+++ internal/tui/list_dialog.go
@@ -1,17 +1,12 @@
 package tui
 
 import (
-	"encoding/binary"
-	"fmt"
-	"hash/fnv"
 	"image/color"
-	"strings"
 
 	"charm.land/bubbles/v2/help"
 	"charm.land/bubbles/v2/key"
 	"charm.land/bubbles/v2/viewport"
 	tea "charm.land/bubbletea/v2"
-	lipgloss "charm.land/lipgloss/v2"
 )
 
 // ListDialogAction is a button rendered in the detail-pane action bar.
@@ -21,37 +16,6 @@ type ListDialogAction struct {
 	Danger   bool
 	Primary  bool
 	Disabled bool
-}
-
-// ListDialogKeys is the minimal key map the shell understands. Callers embed
-// it in their own dialog-specific key map and wire additional hotkeys
-// (e.g. Edit/Delete/RSVP) on top.
-type ListDialogKeys struct {
-	Up       key.Binding
-	Down     key.Binding
-	Tab      key.Binding
-	ShiftTab key.Binding
-	Enter    key.Binding
-	Close    key.Binding
-	PageUp   key.Binding
-	PageDown key.Binding
-	Home     key.Binding
-	End      key.Binding
-}
-
-func defaultListDialogKeys() ListDialogKeys {
-	return ListDialogKeys{
-		Up:       key.NewBinding(key.WithKeys("up", "k"), key.WithHelp("↑/k", "up")),
-		Down:     key.NewBinding(key.WithKeys("down", "j"), key.WithHelp("↓/j", "down")),
-		Tab:      key.NewBinding(key.WithKeys("tab"), key.WithHelp("tab", "sections")),
-		ShiftTab: key.NewBinding(key.WithKeys("shift+tab"), key.WithHelp("shift+tab", "prev section")),
-		Enter:    key.NewBinding(key.WithKeys("enter", "space"), key.WithHelp("enter", "select")),
-		Close:    key.NewBinding(key.WithKeys("esc", "q"), key.WithHelp("esc", "close")),
-		PageUp:   key.NewBinding(key.WithKeys("pgup", "ctrl+b")),
-		PageDown: key.NewBinding(key.WithKeys("pgdown", "ctrl+f")),
-		Home:     key.NewBinding(key.WithKeys("home")),
-		End:      key.NewBinding(key.WithKeys("end")),
-	}
 }
 
 // ListDialogZone identifies the focused region inside the dialog.
@@ -123,19 +87,6 @@ type viewRenderCache struct {
 	actions    string
 	helpKey    uint64
 	help       string
-}
-
-// NewListDialogModel builds an empty shell. Callers call the Setters on the
-// returned value before the render.
-func NewListDialogModel(h help.Model) ListDialogModel {
-	vp := viewport.New()
-	vp.MouseWheelEnabled = true
-	return ListDialogModel{
-		keys:  defaultListDialogKeys(),
-		help:  h,
-		body:  vp,
-		cache: &viewRenderCache{},
-	}
 }
 
 func (m ListDialogModel) SetSize(w, h int) ListDialogModel {
@@ -407,800 +358,3 @@ func (m ListDialogModel) boxSize() (int, int) {
 }
 
 func (m ListDialogModel) isNarrow() bool { return m.width < narrowThreshold }
-
-// MoveUp/MoveDown advance to the next selectable row inside the list zone.
-// They no-op when the list is empty or focus is elsewhere. Selection changes
-// reset the detail viewport so the new row starts from the top.
-func (m ListDialogModel) MoveUp() ListDialogModel {
-	if m.focusZone != ListZoneList {
-		return m
-	}
-	for idx := m.selected - 1; idx >= 0; idx-- {
-		if !m.rowIsDisabled(idx) {
-			m.selected = idx
-			m.body.GotoTop()
-			break
-		}
-	}
-	return m
-}
-
-func (m ListDialogModel) MoveDown() ListDialogModel {
-	if m.focusZone != ListZoneList {
-		return m
-	}
-	for idx := m.selected + 1; idx < len(m.rows); idx++ {
-		if !m.rowIsDisabled(idx) {
-			m.selected = idx
-			m.body.GotoTop()
-			break
-		}
-	}
-	return m
-}
-
-// CycleZone advances focus to the next (or previous) enabled control. Disabled
-// action buttons remain visible but do not participate in the Tab order.
-func (m ListDialogModel) CycleZone(forward bool) ListDialogModel {
-	type focusStop struct {
-		zone   ListDialogZone
-		action int
-	}
-	stops := []focusStop{{zone: ListZoneList}}
-	for idx, action := range m.actions {
-		if !action.Disabled {
-			stops = append(stops, focusStop{zone: ListZoneActions, action: idx})
-		}
-	}
-	if m.titleAction != nil && !m.titleAction.Disabled {
-		stops = append(stops, focusStop{zone: ListZoneTitleAction})
-	}
-	if len(stops) == 1 {
-		m.focusZone = ListZoneList
-		return m
-	}
-
-	current := 0
-	for idx, stop := range stops {
-		if stop.zone == m.focusZone && (stop.zone != ListZoneActions || stop.action == m.focusedAction) {
-			current = idx
-			break
-		}
-	}
-	delta := 1
-	if !forward {
-		delta = -1
-	}
-	next := stops[(current+delta+len(stops))%len(stops)]
-	m.focusZone = next.zone
-	if next.zone == ListZoneActions {
-		m.focusedAction = next.action
-	}
-	return m
-}
-
-// FocusAction focuses an enabled action-bar button.
-func (m ListDialogModel) FocusAction(idx int) ListDialogModel {
-	if idx < 0 || idx >= len(m.actions) || m.actions[idx].Disabled {
-		return m
-	}
-	m.focusZone = ListZoneActions
-	m.focusedAction = idx
-	return m
-}
-
-// ActivateFocused returns the command for whichever zone currently has focus
-// (the focused action button, or the title-action button).
-func (m ListDialogModel) ActivateFocused() tea.Cmd {
-	switch m.focusZone {
-	case ListZoneActions:
-		if m.focusedAction >= 0 && m.focusedAction < len(m.actions) &&
-			!m.actions[m.focusedAction].Disabled {
-			return m.actions[m.focusedAction].Msg
-		}
-	case ListZoneTitleAction:
-		if m.titleAction != nil && !m.titleAction.Disabled {
-			return m.titleAction.Msg
-		}
-	case ListZoneList, ListZoneCustom:
-	}
-	return nil
-}
-
-// RowAtPosition hit-tests a screen-space (x, y) against the rendered list.
-// Returns the row index when the click lands on a row, false otherwise.
-func (m ListDialogModel) RowAtPosition(x, y int) (int, bool) {
-	if len(m.rows) == 0 || m.width <= 0 || m.height <= 0 {
-		return 0, false
-	}
-
-	boxW, boxH := m.boxSize()
-	innerW := max(boxW-5, 10)
-	innerH := max(boxH-3, 6)
-	bodyH := max(innerH-4, 3)
-
-	dialogX := (m.width - boxW) / 2
-	dialogY := (m.height - boxH) / 2
-	listX := dialogX + 2
-	listY := dialogY + 4
-	listW := innerW
-	listH := bodyH
-
-	if m.isNarrow() {
-		listH = min(max(len(m.rows)+1, 3), max(bodyH/3, 3))
-	} else {
-		listW = listColumnWidth(innerW)
-	}
-
-	if x < listX || x >= listX+listW || y < listY || y >= listY+listH {
-		return 0, false
-	}
-
-	row := y - listY
-	if len(m.rows) > listH && row == listH-1 {
-		return 0, false
-	}
-
-	idx := m.scroll + row
-	if idx < 0 || idx >= len(m.rows) || m.rowIsDisabled(idx) {
-		return 0, false
-	}
-	return idx, true
-
-}
-
-// ActionAtPosition hit-tests the action bar. Returns the clicked button index.
-// Each button's width is measured from its actual rendered output. That
-// matches DefaultButtonStyles: Padding(0,2) + MarginRight(1). The hit regions
-// then agree exactly with what the user sees. The join space added by
-// strings.Join(parts, " ") in renderActions accounts for the +1 advance
-// between consecutive buttons.
-func (m ListDialogModel) ActionAtPosition(x, y int) (int, bool) {
-	ox, oy := m.actionBarOrigin()
-	if y != oy {
-		return 0, false
-	}
-	bs := DefaultButtonStyles()
-	cx := ox
-	for i, a := range m.actions {
-		var w int
-		if a.Danger {
-			w = lipgloss.Width(bs.Danger.Render(a.Label, false))
-		} else {
-			w = lipgloss.Width(bs.Normal.Render(a.Label, false))
-		}
-		if x >= cx && x < cx+w {
-			return i, true
-		}
-		cx += w + 1 // +1 for the strings.Join(" ") separator in renderActions
-	}
-	return 0, false
-}
-
-// ClickRow selects idx and focuses the list zone. Resets the detail viewport
-// scroll on selection change so the freshly clicked row's content starts at
-// the top.
-func (m ListDialogModel) ClickRow(idx int) ListDialogModel {
-	if idx < 0 || idx >= len(m.rows) || m.rowIsDisabled(idx) {
-		return m
-	}
-	if idx != m.selected {
-		m.body.GotoTop()
-	}
-	m.selected = idx
-	m.focusZone = ListZoneList
-	return m
-}
-
-// ClickAction focuses an enabled action and returns its command.
-func (m ListDialogModel) ClickAction(idx int) (ListDialogModel, tea.Cmd) {
-	if idx < 0 || idx >= len(m.actions) || m.actions[idx].Disabled {
-		return m, nil
-	}
-	m.focusZone = ListZoneActions
-	m.focusedAction = idx
-	return m, m.actions[idx].Msg
-}
-
-// DetailsOrigin returns the screen-space (x, y) of the first line of the
-// detail pane. Callers can then hit-test buttons they composed into the
-// detail lines (for example RSVP buttons in the event dialog).
-func (m ListDialogModel) DetailsOrigin() (int, int) {
-	boxW, boxH := m.boxSize()
-	dialogX := (m.width - boxW) / 2
-	dialogY := (m.height - boxH) / 2
-	detailsX := dialogX + 2
-	detailsY := dialogY + 4
-	if m.isNarrow() {
-		rowCount := max(len(m.rows), 1)
-		bodyH := max(max(boxH-3, 6)-4, 3)
-		listH := min(max(rowCount+1, 3), max(bodyH/3, 3))
-		detailsY += listH + 1
-	} else {
-		innerW := max(boxW-5, 10)
-		detailsX += listColumnWidth(innerW) + dialogDividerWidth
-	}
-	return detailsX, detailsY
-}
-
-// BodyRowScreenY translates a content-row index inside the scrollable
-// detail body to its screen-space Y. It includes the pinned title row
-// (+2 lines when present) and the current scroll offset. Returns false
-// when the row is scrolled out of view.
-func (m ListDialogModel) BodyRowScreenY(idx int) (int, bool) {
-	_, oy := m.DetailsOrigin()
-	if m.hasPinnedTitle() {
-		oy += 2
-	}
-	visible := idx - m.body.YOffset()
-	if visible < 0 || visible >= m.body.Height() {
-		return 0, false
-	}
-	return oy + visible, true
-}
-
-func (m ListDialogModel) actionBarOrigin() (int, int) {
-	boxW, boxH := m.boxSize()
-	innerW := max(boxW-5, 10)
-	innerH := max(boxH-3, 6)
-	bodyH := max(innerH-4, 3)
-
-	dialogX := (m.width - boxW) / 2
-	dialogY := (m.height - boxH) / 2
-
-	contentX := dialogX + 2
-	actionsY := dialogY + bodyH + 3
-
-	if m.isNarrow() {
-		return contentX, actionsY
-	}
-	return contentX + listColumnWidth(innerW) + dialogDividerWidth, actionsY
-}
-
-// View renders the complete dialog (border, title, body, help row).
-func (m ListDialogModel) View() string {
-	if m.width <= 0 || m.height <= 0 {
-		return ""
-	}
-	boxW, boxH := m.boxSize()
-	innerW := max(boxW-5, 10)
-	innerH := max(boxH-3, 6)
-	bodyH := max(innerH-4, 3)
-
-	title := m.renderTitleRow(innerW)
-	helpText := m.renderHelpLine(innerW)
-
-	var body string
-	if m.isNarrow() {
-		body = m.viewStacked(innerW, bodyH)
-	} else {
-		body = m.viewColumns(innerW, bodyH)
-	}
-
-	// Build the framed dialog manually instead of going through
-	// lipgloss.NewStyle().Border().Render(content): that path forces
-	// lipgloss to re-wrap and re-measure every grapheme of the styled
-	// content, which is by far the biggest single cost on a dense
-	// dialog (96+ rows). Each content line is already innerW cells
-	// wide, so we can splice them between hand-built border + padding
-	// strings and skip the global measurement pass entirely.
-	blank := strings.Repeat(" ", innerW)
-	contentLines := make([]string, 0, innerH)
-	contentLines = append(contentLines, title, m.renderSubtitle(innerW))
-	contentLines = append(contentLines, strings.Split(body, "\n")...)
-
-	contentLines = append(contentLines, blank, helpText)
-	return framedDialog(boxW, contentLines)
-}
-
-// framedDialog wraps innerLines with the rounded border + (1,2,0,1)
-// padding the dialog has always used. innerLines MUST already be at the
-// inner content width (boxW - 5). The title row, helpText, and the
-// row-zipped body are all width-padded by their producers. A skip of
-// the per-line measurement here is then safe. It also saves ~25% of View
-// on a dense dialog because lipgloss.Width is the single biggest cost.
-func framedDialog(boxW int, innerLines []string) string {
-	const (
-		padLeft  = 1
-		padRight = 2
-	)
-	innerW := boxW - 2 - padLeft - padRight
-	top := "╭" + strings.Repeat("─", boxW-2) + "╮"
-	bottom := "╰" + strings.Repeat("─", boxW-2) + "╯"
-	leftPad := strings.Repeat(" ", padLeft)
-	rightPad := strings.Repeat(" ", padRight)
-	emptyRow := "│" + leftPad + strings.Repeat(" ", innerW) + rightPad + "│"
-
-	var b strings.Builder
-	b.Grow(boxW * (len(innerLines) + 3))
-	b.WriteString(top)
-	b.WriteByte('\n')
-	b.WriteString(emptyRow)
-	b.WriteByte('\n')
-	for _, line := range innerLines {
-		b.WriteString("│")
-		b.WriteString(leftPad)
-		b.WriteString(line)
-		b.WriteString(rightPad)
-		b.WriteString("│\n")
-	}
-	b.WriteString(bottom)
-	return b.String()
-}
-
-func (m *ListDialogModel) viewColumns(innerW, bodyH int) string {
-	listW := listColumnWidth(innerW)
-	detailsW := detailColumnWidth(innerW)
-
-	m.adjustScroll(bodyH)
-	// All three column renderers now produce lines that are exactly w
-	// cells wide (renderList via padLines, renderDivider by construction,
-	// renderDetails by padding its variable-width parts internally), so
-	// we can split each without the per-line lipgloss.Width measurement
-	// splitAndPad would do.
-	listCol := trustedSplit(m.renderList(listW, bodyH), listW, bodyH)
-	dividerCol := trustedSplit(m.renderDivider(dialogDividerWidth, bodyH), dialogDividerWidth, bodyH)
-	detailsCol := trustedSplit(m.renderDetails(detailsW, bodyH), detailsW, bodyH)
-
-	// Manual row-zip: lipgloss.JoinHorizontal re-measures every grapheme
-	// of every line in every column to align them. We've already padded
-	// each column to its fixed cell width, so straight concatenation
-	// produces the same visual output without the measurement pass.
-	var b strings.Builder
-	b.Grow((listW + dialogDividerWidth + detailsW + 1) * bodyH)
-	for i := 0; i < bodyH; i++ {
-		if i > 0 {
-			b.WriteByte('\n')
-		}
-		b.WriteString(listCol[i])
-		b.WriteString(dividerCol[i])
-		b.WriteString(detailsCol[i])
-	}
-	return b.String()
-}
-
-// trustedSplit splits s by newlines and returns exactly h rows. Callers
-// must guarantee every emitted line is already w cells wide. Rows that
-// are gone are filled with blanks. It skips lipgloss.Width entirely.
-// That is the whole point of the "trusted" name. The fast path is ~60%
-// faster than splitAndPad on width-correct input.
-func trustedSplit(s string, w, h int) []string {
-	out := make([]string, h)
-	if h <= 0 {
-		return out
-	}
-	lines := strings.Split(s, "\n")
-	n := len(lines)
-	if n > h {
-		n = h
-	}
-	for i := 0; i < n; i++ {
-		out[i] = lines[i]
-	}
-	if n < h {
-		blank := strings.Repeat(" ", w)
-		for i := n; i < h; i++ {
-			out[i] = blank
-		}
-	}
-	return out
-}
-
-func (m *ListDialogModel) viewStacked(innerW, bodyH int) string {
-	rowCount := max(len(m.rows), 1)
-	listH := min(max(rowCount+1, 3), max(bodyH/3, 3))
-	detailsH := max(bodyH-listH-1, 3)
-
-	m.adjustScroll(listH)
-	list := m.renderList(innerW, listH)
-	sep := lipgloss.NewStyle().Faint(true).Width(innerW).
-		Render(strings.Repeat("─", innerW))
-	details := m.renderDetails(innerW, detailsH)
-
-	return lipgloss.JoinVertical(lipgloss.Left, list, sep, details)
-}
-
-func (m *ListDialogModel) adjustScroll(visibleH int) {
-	// When the list overflows, renderList reserves the last visible row
-	// for the scroll indicator (e.g. "5/96 ▼"), overwriting whatever row
-	// would otherwise sit there. Treat that slot as out of bounds for the
-	// selection so the highlighted row never lands on it.
-	contentH := visibleH
-	if len(m.rows) > visibleH && contentH > 1 {
-		contentH = visibleH - 1
-	}
-	if m.selected < m.scroll {
-		m.scroll = m.selected
-	}
-	if m.selected >= m.scroll+contentH {
-		m.scroll = m.selected - contentH + 1
-	}
-	if m.scroll < 0 {
-		m.scroll = 0
-	}
-}
-
-func (m ListDialogModel) renderList(w, h int) string {
-	if len(m.rows) == 0 {
-		if m.emptyList == "" {
-			return padLines(nil, w, h)
-		}
-		msg := lipgloss.NewStyle().Faint(true).Render(m.emptyList)
-		return padLines([]string{msg}, w, h)
-	}
-
-	total := len(m.rows)
-	visibleStart := m.scroll
-	visibleEnd := min(visibleStart+h, total)
-
-	lines := make([]string, 0, h)
-	for i := visibleStart; i < visibleEnd; i++ {
-		lines = append(lines, renderListRow(m.rows[i], w, i == m.selected, m.focusZone == ListZoneList, m.selectedColor))
-	}
-
-	if total > h {
-		indicator := fmt.Sprintf("%d/%d ", m.selected+1, total)
-		arrows := ""
-		if m.scroll > 0 {
-			arrows += "▲"
-		}
-		if visibleEnd < total {
-			if arrows != "" {
-				arrows += " "
-			}
-			arrows += "▼"
-		}
-		if arrows != "" {
-			indicator += arrows + " "
-		}
-		indicator = truncateTo(indicator, w)
-
-		faintIndicator := lipgloss.NewStyle().Faint(true).Render(indicator)
-		if len(lines) >= h {
-			lines[h-1] = faintIndicator
-		} else {
-			lines = append(lines, faintIndicator)
-		}
-	}
-
-	return padLines(lines, w, h)
-}
-
-func (m ListDialogModel) renderDivider(w, h int) string {
-	bar := lipgloss.NewStyle().Faint(true).Render("│")
-	pad := strings.Repeat(" ", (w-1)/2)
-	rest := strings.Repeat(" ", w-len(pad)-1)
-	line := pad + bar + rest
-	lines := make([]string, h)
-	for i := range lines {
-		lines[i] = line
-	}
-	return strings.Join(lines, "\n")
-}
-
-func (m ListDialogModel) renderActions(w int) string {
-	key := m.actionsCacheKey(w)
-	if m.cache != nil && key == m.cache.actionsKey && m.cache.actions != "" {
-		return m.cache.actions
-	}
-	bs := DefaultButtonStyles()
-	parts := make([]string, len(m.actions))
-	for i, a := range m.actions {
-		focused := !a.Disabled && m.focusZone == ListZoneActions && i == m.focusedAction
-		switch {
-		case a.Danger:
-			parts[i] = bs.Danger.Render(a.Label, focused)
-		default:
-			parts[i] = bs.Normal.Render(a.Label, focused)
-		}
-		if a.Disabled {
-			parts[i] = lipgloss.NewStyle().Faint(true).Render(parts[i])
-		}
-	}
-
-	out := truncateTo(strings.Join(parts, " "), w)
-	if m.cache != nil {
-		m.cache.actionsKey = key
-		m.cache.actions = out
-	}
-	return out
-}
-
-// renderHelpLine produces the centered short-help line at the bottom
-// of the dialog. The result is memoized. shortHelp only changes when
-// the caller swaps focus zones or transitions between empty/non-empty
-// states. The cache then skips a full lipgloss render (and the bubbles
-// help layout it wraps) on every key press while the user scrolls.
-func (m ListDialogModel) renderHelpLine(innerW int) string {
-	key := m.helpCacheKey(innerW)
-	if m.cache != nil && key == m.cache.helpKey && m.cache.help != "" {
-		return m.cache.help
-	}
-	m.help.SetWidth(innerW)
-	out := lipgloss.NewStyle().
-		Width(innerW).
-		Align(lipgloss.Center).
-		Render(m.help.ShortHelpView(m.shortHelp))
-	if m.cache != nil {
-		m.cache.helpKey = key
-		m.cache.help = out
-	}
-	return out
-}
-
-// helpCacheKey fingerprints the inputs that affect renderHelpLine: the
-// inner width and every binding's key + help text. shortHelp is
-// rebuilt by the caller on each refresh so identity comparison would
-// always miss. A content-based fingerprint hits whenever the rendered
-// output would be identical.
-func (m ListDialogModel) helpCacheKey(innerW int) uint64 {
-	h := fnv.New64a()
-	var buf [8]byte
-	binary.LittleEndian.PutUint64(buf[:], uint64(innerW))
-	h.Write(buf[:])
-	for _, b := range m.shortHelp {
-		help := b.Help()
-		h.Write([]byte(help.Key))
-		h.Write([]byte{0})
-		h.Write([]byte(help.Desc))
-		h.Write([]byte{0})
-		var flags byte
-		if b.Enabled() {
-			flags = 1
-		}
-		h.Write([]byte{flags})
-	}
-	return h.Sum64()
-}
-
-// actionsCacheKey returns a 64-bit fingerprint of every input that
-// affects renderActions' output. Each Set* on the model that touches
-// one of those inputs naturally changes the fingerprint. The cache
-// then invalidates lazily with no eager records.
-func (m ListDialogModel) actionsCacheKey(w int) uint64 {
-	h := fnv.New64a()
-	var buf [8]byte
-	binary.LittleEndian.PutUint64(buf[:], uint64(w))
-	h.Write(buf[:])
-	binary.LittleEndian.PutUint64(buf[:], uint64(m.focusZone))
-	h.Write(buf[:])
-	binary.LittleEndian.PutUint64(buf[:], uint64(m.focusedAction))
-	h.Write(buf[:])
-	for _, a := range m.actions {
-		h.Write([]byte(a.Label))
-		var flags byte
-		if a.Danger {
-			flags |= 1
-		}
-		if a.Primary {
-			flags |= 2
-		}
-		if a.Disabled {
-			flags |= 4
-		}
-
-		h.Write([]byte{flags, 0})
-	}
-	return h.Sum64()
-}
-
-func (m *ListDialogModel) renderDetails(w, h int) string {
-	lines := m.detailLines
-	if len(m.rows) == 0 {
-		lines = m.emptyDetails
-	}
-
-	bodyH := h
-	var pinned string
-	if m.hasPinnedTitle() {
-		pinned = paneTitle(m.detailTitle, w)
-		bodyH = max(bodyH-2, 1)
-	}
-
-	if len(m.actions) == 0 {
-		m.body.SetWidth(w)
-		m.body.SetHeight(bodyH)
-		m.body.SetContentLines(lines)
-		if pinned != "" {
-			return pinned + "\n" + m.body.View()
-		}
-		return m.body.View()
-	}
-
-	bodyH = max(bodyH-2, 1)
-	m.body.SetWidth(w)
-	m.body.SetHeight(bodyH)
-	m.body.SetContentLines(lines)
-
-	parts := make([]string, 0, 4)
-	if pinned != "" {
-		parts = append(parts, pinned)
-	}
-	// renderActions returns ≤ w cells (truncateTo); pad it so the whole
-	// detail column is width-correct and viewColumns can trustedSplit it
-	// without re-measuring every body line.
-	parts = append(parts, m.body.View(), m.actionsSeparator(w), padTrailing(m.renderActions(w), w))
-	return strings.Join(parts, "\n")
-}
-
-// actionsSeparator renders the faint rule that sits between the detail
-// body and the action bar. When the body has scrolled-away content above
-// or below, a centered "↑↓ more" hint is embedded in the rule to advertise
-// the scroll affordance. That is the same treatment used in the
-// single-event dialog.
-func (m ListDialogModel) actionsSeparator(w int) string {
-	faint := lipgloss.NewStyle().Faint(true)
-	hint := m.scrollHint()
-	hw := lipgloss.Width(hint)
-	if hint == "" || w <= hw+2 {
-		return faint.Render(strings.Repeat("─", w))
-	}
-	left := (w - hw - 2) / 2
-	right := w - hw - 2 - left
-	return faint.Render(strings.Repeat("─", left)) + " " + faint.Render(hint) + " " + faint.Render(strings.Repeat("─", right))
-}
-
-// scrollHint returns "↓ more" / "↑ more" / "↑↓ more" based on what
-// the user can still scroll to. Empty when the body fits with no scroll.
-func (m ListDialogModel) scrollHint() string {
-	if !m.bodyOverflows() {
-		return ""
-	}
-	switch {
-	case m.body.AtTop():
-		return "↓ more"
-	case m.body.AtBottom():
-		return "↑ more"
-	default:
-		return "↑↓ more"
-	}
-}
-
-// bodyOverflows reports whether the detail body has more content than
-// the viewport can show at once.
-func (m ListDialogModel) bodyOverflows() bool {
-	return m.body.TotalLineCount() > m.body.VisibleLineCount()
-}
-
-func (m ListDialogModel) renderSubtitle(innerW int) string {
-	return lipgloss.NewStyle().
-		Faint(true).
-		Width(innerW).
-		Render(truncateTo(m.subtitle, innerW))
-}
-
-// renderTitleRow composes the bold title with optional quiet context and an
-// optional action at the right edge. Context truncates before either control.
-func (m ListDialogModel) renderTitleRow(innerW int) string {
-	var button string
-	if m.titleAction != nil {
-		focused := !m.titleAction.Disabled && m.focusZone == ListZoneTitleAction
-		button = renderTitleActionButton(*m.titleAction, focused)
-	}
-
-	buttonW := lipgloss.Width(button)
-	titleNeed := min(lipgloss.Width(m.title), max(innerW-buttonW, 0))
-	contextBudget := max(innerW-buttonW-titleNeed, 0)
-	context := ""
-	if m.titleContext != "" {
-		separatorW := 2
-		if button != "" {
-			separatorW++
-		}
-		if contextBudget > separatorW {
-			context = truncateTo(m.titleContext, contextBudget-separatorW)
-		}
-	}
-
-	right := button
-	if context != "" {
-		renderedContext := lipgloss.NewStyle().Faint(true).Render(context)
-		right = "  " + renderedContext
-		if button != "" {
-			right += " " + button
-		}
-	}
-	titleW := max(innerW-lipgloss.Width(right), 0)
-	title := lipgloss.NewStyle().
-		Bold(true).
-		Width(titleW).
-		Render(truncateTo(m.title, titleW))
-	return lipgloss.JoinHorizontal(lipgloss.Top, title, right)
-}
-
-// renderTitleActionButton renders a title-line button without the extra
-// margin-right cell used by action-bar buttons. It then sits flush with the
-// dialog's right edge.
-func renderTitleActionButton(a ListDialogAction, focused bool) string {
-	bs := DefaultButtonStyles().Normal
-	style := bs.Normal
-	if focused && !a.Disabled {
-		style = bs.Focused
-	}
-	out := style.UnsetMarginRight().Render(a.Label)
-	if a.Disabled {
-		out = lipgloss.NewStyle().Faint(true).Render(out)
-	}
-	return out
-}
-
-// TitleActionAtPosition reports whether (x, y) lies within the title-line
-// action button and, if so, returns its command.
-func (m ListDialogModel) TitleActionAtPosition(x, y int) (tea.Cmd, bool) {
-	if m.titleAction == nil || m.titleAction.Disabled || m.width <= 0 || m.height <= 0 {
-		return nil, false
-	}
-	boxW, boxH := m.boxSize()
-	innerW := max(boxW-5, 10)
-	dialogX := (m.width - boxW) / 2
-	dialogY := (m.height - boxH) / 2
-	titleY := dialogY + 2
-	if y != titleY {
-		return nil, false
-	}
-	btnW := lipgloss.Width(renderTitleActionButton(*m.titleAction, false))
-	btnStartX := dialogX + 2 + innerW - btnW
-	if x < btnStartX || x >= btnStartX+btnW {
-		return nil, false
-	}
-	return m.titleAction.Msg, true
-}
-
-// HandleKey is the shell's handler for keys it cares about (navigation, tab,
-// enter-on-actions, close). Returns the (maybe-updated) model and the
-// result command. Callers dispatch their domain keys (New/Edit/Delete/…)
-// themselves before they fall through to this.
-func (m ListDialogModel) HandleKey(msg tea.KeyPressMsg, onClose func() tea.Msg) (ListDialogModel, tea.Cmd, bool) {
-	switch {
-	case key.Matches(msg, m.keys.Close):
-		return m, func() tea.Msg { return onClose() }, true
-	case key.Matches(msg, m.keys.Tab):
-		return m.CycleZone(true), nil, true
-	case key.Matches(msg, m.keys.ShiftTab):
-		return m.CycleZone(false), nil, true
-	case key.Matches(msg, m.keys.Up):
-		return m.MoveUp(), nil, true
-	case key.Matches(msg, m.keys.Down):
-		return m.MoveDown(), nil, true
-	case key.Matches(msg, m.keys.PageUp):
-		m.body.PageUp()
-		return m, nil, true
-	case key.Matches(msg, m.keys.PageDown):
-		m.body.PageDown()
-		return m, nil, true
-	case key.Matches(msg, m.keys.Home):
-		m.body.GotoTop()
-		return m, nil, true
-	case key.Matches(msg, m.keys.End):
-		m.body.GotoBottom()
-		return m, nil, true
-	case key.Matches(msg, m.keys.Enter):
-		return m, m.ActivateFocused(), true
-	}
-	return m, nil, false
-}
-
-// HandleMouseWheel forwards mouse wheel events to the detail body. The
-// user can then scroll long event content with the wheel. That is the same
-// affordance the single-event dialog provides.
-func (m ListDialogModel) HandleMouseWheel(msg tea.MouseWheelMsg) (ListDialogModel, tea.Cmd) {
-	var cmd tea.Cmd
-	m.body, cmd = m.body.Update(msg)
-	return m, cmd
-}
-
-// ScrollDetailsUp/Down nudge the detail viewport by one line. Callers use
-// these when up/down arrows belong to the details pane (focus is on actions
-// or RSVP, not on the list itself).
-func (m ListDialogModel) ScrollDetailsUp() ListDialogModel {
-	m.body.ScrollUp(1)
-	return m
-}
-
-func (m ListDialogModel) ScrollDetailsDown() ListDialogModel {
-	m.body.ScrollDown(1)
-	return m
-}
-
-// Keys exposes the shell's default bindings so callers can compose ShortHelp.
-func (m ListDialogModel) Keys() ListDialogKeys { return m.keys }
